requirements.txt
absl-py==0.15.0
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1725434011349/work
aiohttp @ file:///croot/aiohttp_1725527756643/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
astor==0.8.1
astunparse==1.6.3
async-timeout @ file:///croot/async-timeout_1703096998144/work
attrs @ file:///croot/attrs_1729089401488/work
biopython @ file:///home/conda/feedstock_root/build_artifacts/biopython_1704895409820/work
blinker @ file:///croot/blinker_1696539051175/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools @ file:///croot/cachetools_1713977085444/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///croot/click_1698129812380/work
cryptography @ file:///croot/cryptography_1724940562255/work
flatbuffers==1.12
frozenlist @ file:///croot/frozenlist_1698702560391/work
gast==0.3.3
google-auth @ file:///croot/google-auth_1715111149956/work
google-auth-oauthlib==0.4.6
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio==1.32.0
h5py @ file:///tmp/build/80754af9/h5py_1593454122442/work
idna @ file:///croot/idna_1714398848350/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
joblib==1.4.2
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
multidict @ file:///croot/multidict_1701096859099/work
numpy==1.19.5
nvidia-cublas-cu11==11.11.3.6
oauthlib @ file:///croot/oauthlib_1679489621486/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///croot/packaging_1720101850331/work
pandas==1.4.3
platformdirs @ file:///croot/platformdirs_1692205439124/work
pooch @ file:///croot/pooch_1695850093751/work
protobuf==3.20.3
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///croot/pyjwt_1715094751094/work
pyOpenSSL @ file:///croot/pyopenssl_1723651579422/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytz==2025.2
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib @ file:///croot/requests-oauthlib_1720615017878/work
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn==1.3.0
scipy==1.10.1
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==3.5.0
typing-extensions==3.7.4.3
urllib3 @ file:///croot/urllib3_1727769808118/work
Werkzeug @ file:///croot/werkzeug_1716533313107/work
wrapt==1.12.1
yarl @ file:///croot/yarl_1725976495189/work
zipp @ file:///croot/zipp_1729012354496/work
